+++ src/app/components/dashboard/dashboard.component.html
@@ -5,22 +5,23 @@
     </button>
     <img src="../../../assets/images/dashboardIcon.png" class="image">
     <p class="toolbarHeading">MyNotes</p>
-    <mat-form-field class="searchbar" appearance="outline" fxHide.xs="true" fxShow="true">
+
+    <mat-card class="searchbar" fxHide.xs="true" fxShow="true">
+        <mat-icon matPrefix>search</mat-icon>
+        <textarea matInput type="text"  placeholder="Search" class="searchtype" ></textarea>
+    </mat-card>
+
+    <!-- <mat-form-field class="searchbar" appearance="outline" fxHide.xs="true" fxShow="true">
       <mat-icon matPrefix>search</mat-icon>
       <input matInput placeholder="Search">
-    </mat-form-field>
-    <button mat-button class="toolbarButton">
-      <mat-icon >refresh</mat-icon>
-    </button>
-    <button mat-button class="toolbarButton">
-      <mat-icon>list</mat-icon>
-    </button>
-    <button mat-button class="toolbarButton">
-      <mat-icon>settings</mat-icon>
-    </button>
+    </mat-form-field> -->
     <button mat-button class="toolbarButton">
       <mat-icon>apps</mat-icon>
     </button>
+    <button mat-button class="toolbarButton">
+        <mat-icon>apps</mat-icon>
+      </button>
+    
   </mat-toolbar>
 
 <div class="toolnote">
