+++ src/app/components/notes-icons/notes-icons.component.html
@@ -1,3 +1,4 @@
+<div class="container" fxLayoutAlign="space-evenly center" >
 <button mat-icon-button title="Remind me">
     <mat-icon>notifications</mat-icon>
   </button>
@@ -16,3 +17,4 @@
   <button mat-icon-button title="More">
     <mat-icon>more_vert</mat-icon>
   </button>
+</div>
