+++ src/app/components/notes-icons/notes-icons.component.scss
@@ -1,0 +1,4 @@
+.container
+{
+    display: flex;
+}
