+++ src/app/components/notes/notes.component.ts
@@ -6,6 +6,8 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./notes.component.scss']
 })
 export class NotesComponent implements OnInit {
+  
+  show: boolean = true;
 
   constructor() { }
