+++ src/app/components/register/register.component.scss
@@ -5,7 +5,7 @@
   display: flex;
     align-items: center;
     justify-content: center;
-  background-image: url('../../images/loginImageBackgroung.png');
+  background-image: url('../../../assets/images/backgroundImage.png');
 }
 
 .card {
