+++ src/app/components/reset-password/reset-password.component.scss
@@ -4,7 +4,7 @@
   display: flex;
     align-items: center;
     justify-content: center;
-  background-image: url('../../images/loginImageBackgroung.png');
+  background-image: url('../../../assets/images/backgroundImage.png');
 }
 
 .card {
