+++ src/environments/environment.ts
@@ -3,7 +3,8 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  base : 'http://fundoonotes.incubation.bridgelabz.com/api/user/'
 };
 
 /*
